pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.3",
    "inngest>=0.5.18",
    "llama-index-core>=0.14.20",
    "llama-index-readers-file>=0.6.0",
    "openai>=2.32.0",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",    
    "uvicorn>=0.44.0",
]
